Move most things out of main.rs
